+++ src/_h5ai/public/js/lib/model/item.js
@@ -15,7 +15,9 @@ const createLabel = sequence => {
     sequence = sequence.replace(reEndsWithSlash, '');
     try {
         sequence = decodeURIComponent(sequence);
-    } catch (e) {/* skip */ }
+    } catch {
+        /* skip */
+    }
     return sequence;
 };